Added doors and keys, fade in and out, sound effects, music, fixed animator first frame bug.
Finished level 1. Added functionality for hardcore mode.

+++ RDEngine.Shared/Engine/Animation/AnimLayer.cs
@@ -39,7 +39,7 @@ namespace RDEngine.Engine.Animation
         public Animation.VarTypes Type;
         public int VarIndex; //The index of the variable in the corresponding Animator array
 
-        internal int Time;
+        internal float Time;
         internal int CurrKeyFrame; //The index in the array of the current keyframe
 
         private AnimLayer(int index)
